+++ NEWS.md
@@ -1,5 +1,7 @@
 # rxode2 (development version)
 
+- Parameter labels may now be modified via `ini(param = label("text")`
+
 # rxode2 2.0.11
 
 - Give a more meaningful error for 'rxode2' ui models with only error expressions